Changed the pattern used
