--- src/components/StartDatePicker/index.js
@@ -1,3 +1,0 @@
-import StartDatePicker from "./StartDatePicker";
-
-export default StartDatePicker;
